Move framework to use only pandas dataframes after ROOT to Pandas conversion (#86)
* Add skimming to completely convert root files to dataframes 
* Include data frame for reco, evt and gen content.
* Update significance optimisation to work with pure Dataframes
* remove _ML, add new event/run number ID to match candidates with events
* Fix pylint errors

+++ machine_learning_hep/bitwise.py
@@ -51,7 +51,7 @@ def selectbitoff(array_cand_type, bits):
 def filter_bit_df(dfin, namebitmap, activatedbit):
     bitson = activatedbit[0]
     bitsoff = activatedbit[1]
-    array_cand_type = dfin.loc[:, namebitmap].values
+    array_cand_type = dfin.loc[:, namebitmap].values.astype("int")
     res_on = pd.Series([True]*len(array_cand_type))
     res_off = pd.Series([True]*len(array_cand_type))
     res = pd.Series()
